+++ packages/wallet/src/initialization/instances/ramps/index.ts
@@ -1,0 +1,6 @@
+export { rampsService } from './ramps-service/ramps-service';
+export type { RampsServiceInstanceOptions } from './ramps-service/types';
+export { rampsController } from './ramps-controller/ramps-controller';
+export type { RampsControllerInstanceOptions } from './ramps-controller/types';
+export { transakService } from './transak-service/transak-service';
+export type { TransakServiceInstanceOptions } from './transak-service/types';