+++ packages/wallet/src/types.ts
@@ -9,6 +9,9 @@ import type { ApprovalControllerInstanceOptions } from './initialization/instanc
 import type { ConnectivityControllerInstanceOptions } from './initialization/instances/connectivity-controller/types';
 import type { KeyringControllerInstanceOptions } from './initialization/instances/keyring-controller/types';
 import type { NetworkControllerInstanceOptions } from './initialization/instances/network-controller/types';
+import type { RampsControllerInstanceOptions } from './initialization/instances/ramps/ramps-controller/types';
+import type { RampsServiceInstanceOptions } from './initialization/instances/ramps/ramps-service/types';
+import type { TransakServiceInstanceOptions } from './initialization/instances/ramps/transak-service/types';
 import type { RemoteFeatureFlagControllerInstanceOptions } from './initialization/instances/remote-feature-flag-controller/types';
 import type { StorageServiceInstanceOptions } from './initialization/instances/storage-service/types';
 import type { InitializationConfiguration } from './initialization/types';
@@ -28,6 +31,9 @@ export type InstanceSpecificOptions = {
   connectivityController: ConnectivityControllerInstanceOptions;
   keyringController?: KeyringControllerInstanceOptions;
   networkController: NetworkControllerInstanceOptions;
+  rampsController?: RampsControllerInstanceOptions;
+  rampsService?: RampsServiceInstanceOptions;
   remoteFeatureFlagController: RemoteFeatureFlagControllerInstanceOptions;
   storageService: StorageServiceInstanceOptions;
+  transakService?: TransakServiceInstanceOptions;
 };